test: WhitePawn 테스트 추가
