+++ README.md
@@ -37,6 +37,7 @@
       - 폰은 진행 방향으로만 이동할 수 있다.
       - 처음 이동할 때는 최소 한 칸, 최대 두 칸 전진할 수 있다.
       - 적군이 있을 경우, 대각선으로 이동하여 잡을 수 있다.
+      - 기물을 뛰어넘을 수 없다.
 - 도착지에 기물이 있을 경우, 적군이라면 잡을 수 있고 아군이라면 이동할 수 없다.
 
 ### 입력
